rename: [H,h]ttp <- [R,r]emote -- more clarity
